Add files via upload
At last fully uploaded. Some php functions are not included. Logo is downloadable. Plugins are not included. The example of the theme and pest pie i would like to share with wp fans (gooeymenu) can be seen functioning is on the grace.od.ua. Thnak you for visiting my page . Theme files are under MIT license and so you are free to download use and test for own purposes.